+++ prefect_docker/credentials.py
@@ -32,7 +32,7 @@ class DockerRegistryCredentials(Block):
     """
 
     _block_type_name = "Docker Registry Credentials"
-    _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/2IfXXfMq66mrzJBDFFCHTp/6d8f320d9e4fc4393f045673d61ab612/Moby-logo.png?h=250"  # noqa
+    _logo_url = "https://cdn.sanity.io/images/3ugk85nk/production/14a315b79990200db7341e42553e23650b34bb96-250x250.png"  # noqa
     _description = "Store credentials for interacting with a Docker Registry."
 
     username: str = Field(