+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved error message when running Docker worker against an ephemeral API - [#96](https://github.com/PrefectHQ/prefect-docker/pull/96)
+
 ### Deprecated
 
 ### Removed